simplify/modularize detailsTable implementation
